refactor: hoist recipe run-summary formatters into recipes/_logging.py
The science and masters recipes each carried a private _summary() that built
their end-of-run verdict block. Move both into a shared recipes/_logging.py as
science_run_summary() and masters_run_summary() -- one home for the formatters,
recipes stay focused on the reduction flow. Behavior is unchanged; the functions
are moved verbatim and renamed.

- recipes/_logging.py: new module holding both formatters (pure string builders,
  no logging calls -- the recipes pass the result to logger.info).
- Recipes import from recipes._logging. The import is a top-level statement, so
  it runs at exec_module (before main()), failing fast if the recipe is run from
  outside the repo root -- the pipeline's expected working directory.
- Drop the now-unused os import from the science recipe.
- Tests call the formatters directly from recipes._logging instead of reaching
  through the exec-loaded recipe object.
